+++ webapp/app/countryLanding/views/contentCheck/contentCheck.tsx
@@ -3,8 +3,7 @@ import { useDispatch } from 'react-redux'
 
 import { batchActions } from '@webapp/store'
 import { fetchTableData } from '@webapp/components/Assessment/DataTable/actions'
-import { useCountryIso } from '@webapp/hooks'
-
+import { useCountryIso } from '@webapp/store/app'
 import useSectionTables from '@webapp/pages/AssessmentSection/useSectionTables'
 import SectionView from '@webapp/components/Assessment/SectionView'
