+++ webapp/app/countryLanding/views/messageBoard/messageBoardPanel.tsx
@@ -8,7 +8,7 @@ import useI18n from '@webapp/hooks/useI18n'
 import { getRelativeDate } from '@webapp/utils/relativeDate'
 import { profilePictureUri } from '@common/userUtils'
 
-import { useCountryIso } from '@webapp/hooks'
+import { useCountryIso } from '@webapp/store/app'
 import { RootState } from '@webapp/store/RootState'
 import FraReviewFooter from '../../../assessment/components/review/reviewFooter'
