+++ webapp/hooks/useCountryLandingSections.ts
@@ -3,8 +3,7 @@ import React from 'react'
 import { isReviewer } from '@common/countryRole'
 import { isAllowedToChangeRole } from '@common/userManagementAccessControl'
 
-import { useCountryIso } from '@webapp/hooks'
-
+import { useCountryIso } from '@webapp/store/app'
 import MessageBoard from '@webapp/app/countryLanding/views/messageBoard'
 import RecentActivityView from '@webapp/app/countryLanding/views/recentActivityView'
 import LinksView from '@webapp/app/countryLanding/views/linksView'
