+++ webapp/hooks/useIsDataExportView.ts
@@ -1,6 +1,6 @@
 import { Areas } from '@core/country'
 
-import { useCountryIso } from '@webapp/hooks'
+import { useCountryIso } from '@webapp/store/app'
 
 export default () => {
   const countryIso = useCountryIso()
