+++ webapp/store/app/hooks/useCountryIso.ts
@@ -1,6 +1,6 @@
 import { CountryIso } from '@core/country'
 import { useAppSelector } from '@webapp/store'
-import { useIsHome, useIsAdmin } from './useIsPath'
+import { useIsHome, useIsAdmin } from '../../../hooks/useIsPath'
 
 export default (): CountryIso | null => {
   const { countryIso } = useAppSelector((state) => state.app)
