+++ webapp/store/app/index.ts
@@ -7,6 +7,7 @@ export {
   useSecondaryGroupedRegions,
   useAppLoaded,
   usePrintView,
+  useCountryIso,
 } from './hooks'
 
 export { default as AppReducer, AppActions } from './appSlice'
